requirements.txt
python-docx 
pyyaml 
openai
reportlab